Refactors reporting.se.
The dominant change here is making it so `reporting.se` is always in a healthy state.  Previously, there were methods for manipulating the state of this contract that allowed the caller to put this contract into a bad state under the assumption that the caller would follow-up with another call that put the contract back into a good state.  Examples of this were functions like `addRep` and `subtractRep` which would all for more or less than 11,000,000 REP to be in existence.  Now any single method call on this contract will leave this contract in a reasonable and healthy state.  There is no expectation that the caller chains certain calls together to keep the system healthy.

Another change of note is that I have normalized the state in this contract.  REP now is either in `activeRep` or `dormantRep`.  Previously the state was a bit denormalized in that `reputation` contained both `dormant` and `active` rep while `dormant` contained only `dormant`.

Other changes of note:
* I have changed the nomenclature from `fxpamount` to `attorep` to make it clear to the user that we are working with attorep (which equals 10^-18 REP).
* `reportedOnNonFinalRoundTwoEvent` renamed to `reporterLocks` for improved readability and also changed to an array of events, thus allowing a reporter to be locked by multiple events at a time.  I believe this is correct behavior, though I don't fully understand round 2 locking enough to be confident.
* Got rid of the `reporterIndex` concept, reputation is now mapped directly by reporter address.
* Added a number of comments related to bugs, attack vectors and questions.  These should all be resolved shortly, but they can wait until a follow-up PR.
* Removed some duplicate code in `getReputation`
* REP that needs to be redistributed is now stored in `Reputation[branch].dormantRep[0x0000000000000000000000000000000000000000]` rather than branch ID.  This is mainly because branch ID is redundant since we are necessarily already in the branch and this is how it is stored in `init`, so keeps things consistent.
* Removes test faucet support.  We can add this back in temporarily if necessary, though it breaks the contract mentioned in the first paragraph (it puts REP into a state that it shouldn't be allowed to get into by creating REP out of thin air).